begin process of moving MODs to bg_mod.h
