+++ src/shared/bg_misc.cpp
@@ -31,6 +31,7 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 #include "shared/CommonProxies.h"
 #include "bg_public.h"
 #include "parse.h"
+#include "bg_mod.h"
 
 // delayed translation - these strings may be passed to Trans_Gettext() later
 #define N_(x) x
@@ -809,69 +810,6 @@ static void BG_InitMissileAttributes()
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-
-struct meansOfDeathData_t
-{
-	meansOfDeath_t number;
-	const char     *name;
-};
-
-static const meansOfDeathData_t bg_meansOfDeathData[] =
-{
-	{ MOD_UNKNOWN, "MOD_UNKNOWN" },
-	{ MOD_SHOTGUN, "MOD_SHOTGUN" },
-	{ MOD_BLASTER, "MOD_BLASTER" },
-	{ MOD_PAINSAW, "MOD_PAINSAW" },
-	{ MOD_MACHINEGUN, "MOD_MACHINEGUN" },
-	{ MOD_CHAINGUN, "MOD_CHAINGUN" },
-	{ MOD_PRIFLE, "MOD_PRIFLE" },
-	{ MOD_MDRIVER, "MOD_MDRIVER" },
-	{ MOD_LASGUN, "MOD_LASGUN" },
-	{ MOD_LCANNON, "MOD_LCANNON" },
-	{ MOD_LCANNON_SPLASH, "MOD_LCANNON_SPLASH" },
-	{ MOD_FLAMER, "MOD_FLAMER" },
-	{ MOD_FLAMER_SPLASH, "MOD_FLAMER_SPLASH" },
-	{ MOD_BURN, "MOD_BURN" },
-	{ MOD_GRENADE, "MOD_GRENADE" },
-	{ MOD_FIREBOMB, "MOD_FIREBOMB" },
-	{ MOD_WEIGHT_H, "MOD_WEIGHT_H" },
-	{ MOD_WATER, "MOD_WATER" },
-	{ MOD_SLIME, "MOD_SLIME" },
-	{ MOD_LAVA, "MOD_LAVA" },
-	{ MOD_CRUSH, "MOD_CRUSH" },
-	{ MOD_TELEFRAG, "MOD_TELEFRAG" },
-	{ MOD_FALLING, "MOD_FALLING" },
-	{ MOD_SUICIDE, "MOD_SUICIDE" },
-	{ MOD_TARGET_LASER, "MOD_TARGET_LASER" },
-	{ MOD_TRIGGER_HURT, "MOD_TRIGGER_HURT" },
-	{ MOD_ABUILDER_CLAW, "MOD_ABUILDER_CLAW" },
-	{ MOD_LEVEL0_BITE, "MOD_LEVEL0_BITE" },
-	{ MOD_LEVEL1_CLAW, "MOD_LEVEL1_CLAW" },
-	{ MOD_LEVEL3_CLAW, "MOD_LEVEL3_CLAW" },
-	{ MOD_LEVEL3_POUNCE, "MOD_LEVEL3_POUNCE" },
-	{ MOD_LEVEL3_BOUNCEBALL, "MOD_LEVEL3_BOUNCEBALL" },
-	{ MOD_LEVEL2_CLAW, "MOD_LEVEL2_CLAW" },
-	{ MOD_LEVEL2_ZAP, "MOD_LEVEL2_ZAP" },
-	{ MOD_LEVEL4_CLAW, "MOD_LEVEL4_CLAW" },
-	{ MOD_LEVEL4_TRAMPLE, "MOD_LEVEL4_TRAMPLE" },
-	{ MOD_WEIGHT_A, "MOD_WEIGHT_A" },
-	{ MOD_SLOWBLOB, "MOD_SLOWBLOB" },
-	{ MOD_POISON, "MOD_POISON" },
-	{ MOD_SWARM, "MOD_SWARM" },
-	{ MOD_HSPAWN, "MOD_HSPAWN" },
-	{ MOD_ROCKETPOD, "MOD_ROCKETPOD" },
-	{ MOD_MGTURRET, "MOD_MGTURRET" },
-	{ MOD_REACTOR, "MOD_REACTOR" },
-	{ MOD_ASPAWN, "MOD_ASPAWN" },
-	{ MOD_ATUBE, "MOD_ATUBE" },
-	{ MOD_SPIKER, "MOD_SPIKER" },
-	{ MOD_OVERMIND, "MOD_OVERMIND" },
-	{ MOD_DECONSTRUCT, "MOD_DECONSTRUCT" },
-	{ MOD_REPLACE, "MOD_REPLACE" },
-	{ MOD_BUILDLOG_REVERT, "MOD_BUILDLOG_REVERT" },
-};
-
-static const size_t bg_numMeansOfDeath = ARRAY_LEN( bg_meansOfDeathData );
 
 /*
 ==============
